+++ CHANGELOG.md
@@ -257,6 +257,7 @@ Orita 第一个正式版本。
 
 - Supported `MSVC`.
 - Add Github Actions.
+- Supported local settings.
 
 #### Fixes